pyproject.toml
[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
isort = "^6.0.1"
mypy = "^1.15.0"
pytest = "^8.3.5"
pre-commit = "^4.2.0"
types-pillow = "^10.2.0.20240822"

[tool.poetry]
packages = [
  { include = "trading_cards", from = "src" }
]

[project]
name = "trading-cards-2025"
version = "0.1.0"
description = "A trading card generator script for Camp Mivoden staff trading cards."
authors = [
    {name = "Luke Irvine",email = "lukeirvinefreelance@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "watchdog (>=6.0.0,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
